change T/R line to use one of the lines in the TD62783
This is because the furuno LPF needs a positive value and has no pull
up at the input. The TD62783 is ideal because it has an internal pull
up and can directly drive one of the lines. So, the 8th line (last
one) is used as T/R drive pin.