revamp NR command handling
This changes the behaviour of NR/NR2 (ZZNR/ZZNS in powersdr command
set) and instead consolidates everything into ZZNR. Sending "ZZNR;"
returns a number between 0..4, both inclusive. 0 means, NR is off, 1-4
indicates one of NR/NR2/NR3/NR4 is active. Similarly to set a
particular NR, send "ZZNRx;" where x is a number between 0-4.

A corresponding pihpsdr change also needs to be picked up for this to
work, so pull in the pihpsdr and build/install it as well.